+++ sm2/tests/sm2pke.rs
@@ -68,21 +68,21 @@ prop_compose! {
 
 proptest! {
     #[test]
-    fn encrypt_and_decrpyt_der(dk in decrypting_key()) {
+    fn encrypt_and_decrypt_der(dk in decrypting_key()) {
         let ek = dk.encrypting_key();
         let cipher_bytes = ek.encrypt_der(&mut OsRng, MSG).unwrap();
         prop_assert!(dk.decrypt_der(&cipher_bytes).is_ok());
     }
 
     #[test]
-    fn encrypt_and_decrpyt(dk in decrypting_key()) {
+    fn encrypt_and_decrypt(dk in decrypting_key()) {
         let ek = dk.encrypting_key();
         let cipher_bytes = ek.encrypt(&mut OsRng, MSG).unwrap();
         assert_eq!(dk.decrypt(&cipher_bytes).unwrap(), MSG);
     }
 
     #[test]
-    fn encrypt_and_decrpyt_mode(dk in decrypting_key_c1c2c3()) {
+    fn encrypt_and_decrypt_mode(dk in decrypting_key_c1c2c3()) {
         let ek = dk.encrypting_key();
         let cipher_bytes = ek.encrypt(&mut OsRng, MSG).unwrap();
         assert_eq!(